Leyak Paperwork Buff + windoor (#5277)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I changed the trashcan in the bridge for a photocopier and made the
armoury windoor rotated inward and put a public one on the outside so
you can put stuff on the table like the HOP line

<img width="325" height="319" alt="image"
src="https://github.com/user-attachments/assets/ca775cee-8707-4ae7-a944-6496bf62492d"
/>

<img width="319" height="339" alt="image"
src="https://github.com/user-attachments/assets/6c1b65b8-c429-470e-a4f9-da8cf870eda3"
/>


## Why It's Good For The Game
photocopiers are like godsgrace to the land
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: The Leyak has a photocopier now. It also has it's indoor adjusted.

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->