Adds a few modified meson goggles to the Derecho (#5301)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a modified meson goggle to the captain's locker, the
foreman's locker, and the wrecker's lockers on the Derecho.

## Why It's Good For The Game

This should allow for additional outfit customization options aboard the
ship.

## Changelog

:cl:
add: Added a few modified mesons to the Derecho
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->