Adds claymores to the outpost export market, and removes them from outpost cargo (#5302)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds claymores as a possible export at the outpost market, at
1250 credits each as the base price, and it removes the option to
purchase claymores on the market.

## Why It's Good For The Game

No reward without a little risk, right? I remember seeing someone taking
the time to carefully defuse a few in order to sell them, only for them
to wind up disappointed because they weren't actually a valid export. It
makes sense for CLIP to want to get these removed from the sector as
much as landmines, I believe, so this PR aims to add that option.

Claymores as they currently are cost too much to be useful for a
defensive emplacement, and they're also just a little too destructive
when they _do_ catch someone, oftentimes instantly killing someone and
shredding much of their equipment.

## Changelog

:cl:
add: Added claymore exports
del: Removed claymore purchases
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->