Gives the Mission Network app icons on the Laptop & Modular Tablet (#5335)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Gives the Mission Network app icons on the Laptop & Modular Tablet,
which were referenced in the PR that added the apps, but did not exist.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Flavor, plus reminds you that you left the laptop on
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl: riquewright
imageadd: added missing Mission Network App icon states to Laptop &
Tablet
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

![yellow](https://github.com/user-attachments/assets/d040175b-c7d3-4808-a2de-2091c5562c56)