Kali Andhi tweaks (#5358)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds a couple of things to the Kali Andhi, fixes a minor bug,
and tweaks one other thing
- Goggles are added to the Captain, LT and Ensign's lockers
- The Captain, LT, Ensign, and Operative's lockers are given a security
messenger bag in addition to the standard backpack that comes in their
lockers (is this appropriate? I'm not sure -- I can instead swap them
with a plain messenger bag if not)
- The mechanic's locker is given a standard grinder pack and a dufflebag
- The portside and starboard side cargo buttons on the table in the
cargo bay didn't work before, but they should now
- The railing in the cargo bay surrounding the oxygen canister in the
bay had its layer adjusted so that the railing now renders under the
canister instead of over it
- A few missing pipes to the combustion thrusters were added

Before:
<img width="130" height="121" alt="image"
src="https://github.com/user-attachments/assets/2b65e825-4a7e-4602-8679-f9dd68a3794d"
/>

After:
<img width="131" height="140" alt="image"
src="https://github.com/user-attachments/assets/418c18f0-a2d5-4c1b-8616-23d851256b8e"
/>

## Why It's Good For The Game

The aim of this PR is to add a bit more flexibility to players in order
to select the outfit they want to present, and to bring a bit of parity
with the previous Kali Andhi iteration.

## Changelog

:cl:
add: Added a bunch of equipment to the Kali Andhi
fix: fixed a few buttons and a mismatched layer
fix: fixed missing pipes to the combustion engines in the Kali Andhi
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->