Renames NGR Initiates to Crewmen (#5398)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR renames Initiates of NGR ships so as to reflavor them as
Crewmen.

## Why It's Good For The Game

Much like how renaming Recruits for the IRMG to Auxiliaries helped open
up their usage much more, I believe that the same thing can be done with
Initiates of the NGR.

## Changelog

:cl:
add: Reflavored Initiates to be Crewmen
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->