Fixes the derecho smelter console so that it links with the furnace (#5414)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This is a minor fix to the Derecho's smelter console that sets the dist
variable from its default of 1, to 3. This enables it to see the
smelter.

## Why It's Good For The Game

The console would end up unlinked to the smelter in the Derecho, and
sometimes manually linking the two together with a multitool would fail.
This PR should address the first issue.

## Changelog

:cl:
fix: fixed the Derecho's smelter console so that it links to the smelter
at init
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->