Slightly increases the effectiveness of partial understanding of languages (#5474)
## About The Pull Request

The chance of understanding a word is now +10% to -5% based on how
common the word is, instead of +10% to -15%.

## Why It's Good For The Game

Currently, having a very high but not complete understanding of a
language is far less effective than intended because of the -15% penalty
the majority of words will have. This is especially noticeable with the
mutual intelligibility between Gezenan and Zohilan, where it more than
doubles the chance you will not be able to understand a given word.

## Changelog

:cl:
balance: Slightly increased the effectiveness of partial understanding
of languages
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->