Fixes metal dupe with surplus (#5480)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A base-level autolathe can print out surplus ammo boxes for 12 sheets of
metal. These boxes come with four ammo handfuls, which can each be
recycled in an autolathe to produce 7.5 units of metal. With all four
handfuls, a total of 30 sheets of metal can be produced.

This PR changes the cost of both surplus ammo boxes, and the value you
can get from recycling the handfuls they hold. Surplus ammo now costs a
little more than 12 sheets of metal to print, and each handful can now
only be recycled for 2 sheets of metal. With a fully upgraded autolathe,
the cost to print surplus ammunition boxes drops to 8 sheets per box.

## Why It's Good For The Game

It's a little silly that you can get more from less out of thin air, but
I suppose a lot of things in the frontier are like that.

## Changelog

:cl:
fix: fixes a metal dupe exploit
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->