Fixes a missing icon state for the ammotype selector for the Kalixcian Volleygun (#5413)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The Etherbor Volleygun uses the "kill" and "disable" modes, but a kill
icon was not made at the time -- this should rectify the issue.

<img width="221" height="218" alt="image"
src="https://github.com/user-attachments/assets/efec7522-22d2-40bd-9b3d-57fce2a84f9e"
/>

## Why It's Good For The Game

This is just a minor fix for a mostly unremarkable bug.

## Changelog

:cl:
fix: fixed a missing ammotype iconstate for the VG line of etherbor
weapons
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: Erika Fox <94164348+Erikafox@users.noreply.github.com>