Puts a command console in shetland bridge (#5491)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Moves some shit around in the shetland to put a command console

wideband moved to where the recharger was, recharger moved to where the
folders/pens were, folders/pens put in the filing cabinet

<img width="409" height="463" alt="image"
src="https://github.com/user-attachments/assets/8c8195f5-7c3d-4096-985c-41bbc7db2dbb"
/>

## Why It's Good For The Game
Not having a command console in the bridge is just annoying and doesn't
add anything to the ship. command needs a console to command things and
having to walk all the way down to the ground floor to claim stuff is
just annoying.

if it's not 'in spirit' of the shetland i can just replace it with a
laptop that has an id module (or even a tablet on a table with a id
module if we want real lemon ass ship)
## Changelog

:cl: Goat
add: Shetland now has a command console in its bridge
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->